+++ src/action.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 import { ActionNoAckDef, ActionWithAckDef } from "./actions-factory";
 import { EmissionMap } from "./emission";
 import { ActionContext, ClientContext, Handler } from "./handler";
-import { Namespaces, rootNS } from "./namespaces";
+import { Namespaces, rootNS } from "./namespace";
 
 export abstract class AbstractAction {
   public abstract getEvent(): string;
